# docs/openapi/smartranking-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/players": {
      "post": {
        "operationId": "Players_createPlayer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      },
      "get": {
        "operationId": "Players_getAllPlayers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search on name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": false,
            "in": "query",
            "description": "Filter by club (system admins only)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/players/search": {
      "get": {
        "operationId": "Players_searchPlayers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search on name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": false,
            "in": "query",
            "description": "Filter by club (system admins only)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/players/by-email/{email}": {
      "get": {
        "operationId": "Players_getPlayerByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/players/by-club/{clubId}": {
      "get": {
        "operationId": "Players_getPlayersByClub",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/players/by-phone": {
      "get": {
        "operationId": "Players_getPlayerByPhone",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/players/{_id}": {
      "get": {
        "operationId": "Players_getPlayerByid",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      },
      "put": {
        "operationId": "Players_updatePlayer",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      },
      "delete": {
        "operationId": "Players_deletePlayer",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Players"
        ]
      }
    },
...
    "/api/v1/categories/{category}/players/{playerId}": {
      "post": {
        "operationId": "Categories_assignPlayerCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/api/v1/matches": {
      "post": {
        "operationId": "Matches_createMatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Matches"
        ]
      },
      "get": {
        "operationId": "Matches_getMatches",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "clubId",
            "required": false,
            "in": "query",
            "description": "Restrict matches to a specific club (system admins only)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Restrict matches to a single category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive ISO 8601 start date for playedAt (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive ISO 8601 end date for playedAt (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/v1/matches/by-category/{categoryId}": {
      "get": {
        "operationId": "Matches_getMatchesByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive ISO 8601 start date for playedAt (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive ISO 8601 end date for playedAt (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/v1/matches/ranking/{categoryId}": {
      "get": {
        "operationId": "Matches_getRankingByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Optional search over player name or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Matches"
        ]
      }
    }
  },
  "info": {
    "title": "SmartRanking API",
    "description": "Multi-tenant tennis ranking API with Better Auth, RBAC, and MongoDB.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Better Auth session token set via /api/auth"
      },
      "Tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id",
        "description": "Tenant/club identifier. Required for all protected multi-tenant endpoints."
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePlayerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePlayerDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateMatchDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}